refactor : Game Page 관심사 분리 * components 파일 생성 * Content, Header 파일 생성
